refactor(prompt): move SelectOne to select.go
